// src/interfaces/ICharacterCard.tsx
interface ICharacterCard {
  id: number;
  name: string;
  status: string;
  species: string;
  gender: string;
  type: string;
  origin: {
    name: string;
    url: string;
  };
  location: {
    name: string;
    url: string;
  };
  image: string;
  episode: string[];
  url: string;
  created: string;
}

export type { ICharacterCard };